// types/interval.ts
export type Interval =
  | 'minor 2nd'
  | 'major 2nd'
  | 'minor 3rd'
  | 'major 3rd'
  | 'perfect 4th'
  | 'tritone'
  | 'augmented 4th'
  | 'diminished 5th'
  | 'perfect 5th'
  | 'minor 6th'
  | 'augmented 5th'
  | 'major 6th'
  | 'minor 7th'
  | 'augmented 6th'
  | 'major 7th'

export const EVERY_INTERVALS: Interval[] = [
  'minor 2nd',
  'major 2nd',
  'minor 3rd',
  'major 3rd',
  'perfect 4th',
  'tritone',
  'augmented 4th',
  'diminished 5th',
  'perfect 5th',
  'minor 6th',
  'augmented 5th',
  'major 6th',
  'minor 7th',
  'augmented 6th',
  'major 7th',
]